Move search_and_replace to a tool file
